@@ -3,6 +3,8 @@ import 'package:provider/provider.dart';
 
 import '../l10n/app_localizations.dart';
 import '../viewmodels/app_settings_viewmodel.dart';
+import '../widgets/accessibility_widgets.dart';
+import '../widgets/error_boundary.dart';
 
 class SettingsScreen extends StatelessWidget {
   static const route = '/settings';
@@ -12,39 +14,45 @@ class SettingsScreen extends StatelessWidget {
   Widget build(BuildContext context) {
     final s = AppLocalizations.of(context)!;
     final vm = context.watch<AppSettingsViewModel>();
-    return Scaffold(
-      appBar: AppBar(title: Text(s.settings)),
-      body: ListView(
-        padding: const EdgeInsets.all(16),
-        children: [
-          SwitchListTile(
-            value: vm.highContrast,
-            onChanged: vm.toggleHighContrast,
-            title: Text(s.highContrast),
-          ),
-          ListTile(
-            title: Text(s.language),
-            subtitle: Text(vm.locale?.toLanguageTag() ?? 'system'),
-            onTap: () async {
-              final locale = await showDialog<Locale?>(
-                context: context,
-                builder: (_) => const _LanguageDialog(),
-              );
-              if (locale != null) vm.setLocale(locale);
-            },
-          ),
-          ListTile(
-            title: const Text('Text size'),
-            subtitle: Slider(
-              value: vm.textScale,
-              onChanged: vm.setTextScale,
-              min: 0.8,
-              max: 2.0,
-              divisions: 12,
-              label: '${(vm.textScale * 100).toStringAsFixed(0)}%',
+    return ErrorBoundary(
+      context: 'SettingsScreen',
+      onRetry: () {
+        // Retry functionality
+      },
+      child: Scaffold(
+        appBar: AppBar(title: AccessibleText(s.settings)),
+        body: ListView(
+          padding: const EdgeInsets.all(16),
+          children: [
+            SwitchListTile(
+              value: vm.highContrast,
+              onChanged: vm.toggleHighContrast,
+              title: AccessibleText(s.highContrast),
             ),
-          ),
-        ],
+            ListTile(
+              title: AccessibleText(s.language),
+              subtitle: AccessibleText(vm.locale?.toLanguageTag() ?? 'system'),
+              onTap: () async {
+                final locale = await showDialog<Locale?>(
+                  context: context,
+                  builder: (_) => const _LanguageDialog(),
+                );
+                if (locale != null) vm.setLocale(locale);
+              },
+            ),
+            ListTile(
+              title: const AccessibleText('Text size'),
+              subtitle: Slider(
+                value: vm.textScale,
+                onChanged: vm.setTextScale,
+                min: 0.8,
+                max: 2.0,
+                divisions: 12,
+                label: '${(vm.textScale * 100).toStringAsFixed(0)}%',
+              ),
+            ),
+          ],
+        ),
       ),
     );
   }
@@ -56,20 +64,20 @@ class _LanguageDialog extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return AlertDialog(
-      title: const Text('Choose language'),
+      title: const AccessibleText('Choose language'),
       content: Column(
         mainAxisSize: MainAxisSize.min,
         children: [
           ListTile(
-            title: const Text('English (AU)'),
+            title: const AccessibleText('English (AU)'),
             onTap: () => Navigator.pop(context, const Locale('en', 'AU')),
           ),
           ListTile(
-            title: const Text('Yolngu (demo)'),
+            title: const AccessibleText('Yolngu (demo)'),
             onTap: () => Navigator.pop(context, const Locale.fromSubtags(languageCode: 'yol')),
           ),
           ListTile(
-            title: const Text('System default'),
+            title: const AccessibleText('System default'),
             onTap: () => Navigator.pop(context, null),
           ),
         ],
